backend/ofa/requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cors==1.0.1
cycler==0.12.1
distlib==0.3.8
Django==4.2.11
django-bootstrap-form==3.4
django-bootstrap-v5==1.0.11
django-filter==24.2
filelock==3.12.4
Flask==3.0.0
Flask-APScheduler==1.13.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.49.0
future==0.18.3
gevent==23.9.1
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.21
sqlparse==0.4.4
tldextract==3.6.0
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.1
urllib3==2.0.6
virtualenv==20.25.1
Werkzeug==3.0.0
whitenoise==6.6.0
zope.event==5.0
zope.interface==6.0
